Rename the data.js file to BookElement
